レコメンドのアルゴリズムの実装

+++ model/tag.go
@@ -3,5 +3,5 @@ package model
 type Tag struct {
 	ID         uint64 `db:"id"`
 	TagName    string `db:"tag_name"`
-	CategoryID uint64 `db:"category_id"`
+	CategoryID int    `db:"category_id"`
 }
